adjust test structure
