+++ jvm-packages/pom.xml
@@ -25,16 +25,6 @@
     </modules>
     <profiles>
         <profile>
-            <id>spark-1.x</id>
-            <activation>
-                <activeByDefault>false</activeByDefault>
-            </activation>
-            <properties>
-                <spark.version>1.6.1</spark.version>
-                <scala.binary.version>2.10</scala.binary.version>
-            </properties>
-        </profile>
-        <profile>
             <id>spark-2.x</id>
             <activation>
                 <activeByDefault>true</activeByDefault>
