+++ jvm-packages/xgboost4j-spark/src/test/scala/ml/dmlc/xgboost4j/scala/spark/XGBoostGeneralSuite.scala
@@ -19,138 +19,17 @@ package ml.dmlc.xgboost4j.scala.spark
 import java.io.File
 import java.nio.file.Files
 
-import scala.collection.mutable
 import scala.collection.mutable.ListBuffer
-import scala.io.Source
 import scala.util.Random
 
-import ml.dmlc.xgboost4j.scala.spark.DataUtils._
-import org.apache.commons.logging.LogFactory
-import org.apache.spark.mllib.linalg.{Vector => SparkVector, VectorUDT, Vectors, DenseVector}
+import ml.dmlc.xgboost4j.java.{Booster => JBooster, DMatrix => JDMatrix}
+import ml.dmlc.xgboost4j.scala.{Booster, DMatrix, XGBoost => ScalaXGBoost}
+import org.apache.spark.mllib.linalg.{Vector => SparkVector, Vectors}
 import org.apache.spark.mllib.regression.LabeledPoint
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.types._
-import org.apache.spark.sql.{SparkSession, Row, DataFrame}
 import org.apache.spark.{SparkConf, SparkContext}
-import org.scalatest.{BeforeAndAfter, FunSuite}
 
-import ml.dmlc.xgboost4j.java.{Booster => JBooster, DMatrix => JDMatrix, XGBoostError}
-import ml.dmlc.xgboost4j.scala.{Booster, DMatrix, EvalTrait, XGBoost => ScalaXGBoost}
-
-class XGBoostSuite extends FunSuite with BeforeAndAfter {
-
-  private implicit var sc: SparkContext = null
-  private val numWorkers = Runtime.getRuntime().availableProcessors()
-
-  private class EvalError extends EvalTrait {
-
-    val logger = LogFactory.getLog(classOf[EvalError])
-
-    private[xgboost4j] var evalMetric: String = "custom_error"
-
-    /**
-     * get evaluate metric
-     *
-     * @return evalMetric
-     */
-    override def getMetric: String = evalMetric
-
-    /**
-     * evaluate with predicts and data
-     *
-     * @param predicts predictions as array
-     * @param dmat     data matrix to evaluate
-     * @return result of the metric
-     */
-    override def eval(predicts: Array[Array[Float]], dmat: DMatrix): Float = {
-      var error: Float = 0f
-      var labels: Array[Float] = null
-      try {
-        labels = dmat.getLabel
-      } catch {
-        case ex: XGBoostError =>
-          logger.error(ex)
-          return -1f
-      }
-      val nrow: Int = predicts.length
-      for (i <- 0 until nrow) {
-        if (labels(i) == 0.0 && predicts(i)(0) > 0) {
-          error += 1
-        } else if (labels(i) == 1.0 && predicts(i)(0) <= 0) {
-          error += 1
-        }
-      }
-      error / labels.length
-    }
-  }
-
-  before {
-    // build SparkContext
-    val sparkConf = new SparkConf().setMaster("local[*]").setAppName("XGBoostSuite")
-    sc = new SparkContext(sparkConf)
-    sc.setLogLevel("ERROR")
-  }
-
-  after {
-    if (sc != null) {
-      sc.stop()
-    }
-  }
-
-  private def loadLabelPoints(filePath: String): List[LabeledPoint] = {
-    val file = Source.fromFile(new File(filePath))
-    val sampleList = new ListBuffer[LabeledPoint]
-    for (sample <- file.getLines()) {
-      sampleList += fromSVMStringToLabeledPoint(sample)
-    }
-    sampleList.toList
-  }
-
-  private def loadRow(filePath: String): List[Row] = {
-    val file = Source.fromFile(new File(filePath))
-    val rowList = new ListBuffer[Row]
-    for (rowLine <- file.getLines()) {
-      rowList += fromSVMStringToRow(rowLine)
-    }
-    rowList.toList
-  }
-
-  private def fromSVMStringToLabelAndVector(line: String): (Double, SparkVector) = {
-    val labelAndFeatures = line.split(" ")
-    val label = labelAndFeatures(0).toDouble
-    val features = labelAndFeatures.tail
-    val denseFeature = new Array[Double](129)
-    for (feature <- features) {
-      val idAndValue = feature.split(":")
-      denseFeature(idAndValue(0).toInt) = idAndValue(1).toDouble
-    }
-    (label, new DenseVector(denseFeature))
-  }
-
-  private def fromSVMStringToLabeledPoint(line: String): LabeledPoint = {
-    val (label, sv) = fromSVMStringToLabelAndVector(line)
-    LabeledPoint(label, sv)
-  }
-
-  private def fromSVMStringToRow(line: String): Row = {
-    val (label, sv) = fromSVMStringToLabelAndVector(line)
-    Row(label, sv)
-  }
-
-  private def buildTrainingDataframe(sparkContext: Option[SparkContext] = None):
-      DataFrame = {
-    val rowList = loadRow(getClass.getResource("/agaricus.txt.train").getFile)
-    val rowRDD = sparkContext.getOrElse(sc).parallelize(rowList, numWorkers)
-    val sparkSession = SparkSession.builder().appName("XGBoostTest").getOrCreate()
-    sparkSession.createDataFrame(rowRDD,
-      StructType(Array(StructField("label", DoubleType, nullable = false),
-        StructField("features", new VectorUDT, nullable = false))))
-  }
-
-  private def buildTrainingRDD(sparkContext: Option[SparkContext] = None): RDD[LabeledPoint] = {
-    val sampleList = loadLabelPoints(getClass.getResource("/agaricus.txt.train").getFile)
-    sparkContext.getOrElse(sc).parallelize(sampleList, numWorkers)
-  }
+class XGBoostGeneralSuite extends Utils {
 
   test("build RDD containing boosters with the specified worker number") {
     val trainingRDD = buildTrainingRDD()
@@ -196,77 +75,6 @@ class XGBoostSuite extends FunSuite with BeforeAndAfter {
     val dir = new File(".")
     for (file <- dir.listFiles() if file.getName.startsWith("XGBoostSuite-0-dtrain_cache")) {
       file.delete()
-    }
-  }
-
-  test("test consistency between training with dataframe and RDD") {
-    val trainingDF = buildTrainingDataframe()
-    val trainingRDD = buildTrainingRDD()
-    val paramMap = List("eta" -> "1", "max_depth" -> "6", "silent" -> "0",
-      "objective" -> "binary:logistic").toMap
-    val xgBoostModelWithDF = XGBoost.trainWithDataset(trainingDF, paramMap,
-      round = 5, nWorkers = numWorkers, useExternalMemory = false)
-    val xgBoostModelWithRDD = XGBoost.trainWithRDD(trainingRDD, paramMap,
-      round = 5, nWorkers = numWorkers, useExternalMemory = false)
-    val eval = new EvalError()
-    val testSet = loadLabelPoints(getClass.getResource("/agaricus.txt.test").getFile).iterator
-    import DataUtils._
-    val testSetDMatrix = new DMatrix(new JDMatrix(testSet, null))
-    assert(
-      eval.eval(xgBoostModelWithDF.booster.predict(testSetDMatrix, outPutMargin = true),
-        testSetDMatrix) ===
-      eval.eval(xgBoostModelWithRDD.booster.predict(testSetDMatrix, outPutMargin = true),
-        testSetDMatrix))
-  }
-
-  test("test transform of dataframe-based model") {
-    val trainingDF = buildTrainingDataframe()
-    val paramMap = List("eta" -> "1", "max_depth" -> "6", "silent" -> "0",
-      "objective" -> "binary:logistic").toMap
-    val xgBoostModelWithDF = XGBoost.trainWithDataset(trainingDF, paramMap,
-      round = 5, nWorkers = numWorkers, useExternalMemory = false)
-    val testSet = loadLabelPoints(getClass.getResource("/agaricus.txt.test").getFile)
-    val testRowsRDD = sc.parallelize(testSet.zipWithIndex, numWorkers).map{
-      case (instance: LabeledPoint, id: Int) =>
-        Row(id, instance.features, instance.label)
-    }
-    val testDF = trainingDF.sparkSession.createDataFrame(testRowsRDD, StructType(
-      Array(StructField("id", IntegerType),
-        StructField("features", new VectorUDT), StructField("label", DoubleType))))
-    xgBoostModelWithDF.transform(testDF).show()
-  }
-
-  test("test order preservation of dataframe-based model") {
-    val paramMap = List("eta" -> "1", "max_depth" -> "6", "silent" -> "0",
-      "objective" -> "binary:logistic").toMap
-    val trainingItr = loadLabelPoints(getClass.getResource("/agaricus.txt.train").getFile).
-      iterator
-    val (testItr, auxTestItr) =
-      loadLabelPoints(getClass.getResource("/agaricus.txt.test").getFile).iterator.duplicate
-    import DataUtils._
-    val trainDMatrix = new DMatrix(new JDMatrix(trainingItr, null))
-    val testDMatrix = new DMatrix(new JDMatrix(testItr, null))
-    val xgboostModel = ScalaXGBoost.train(trainDMatrix, paramMap, 5)
-    val predResultFromSeq = xgboostModel.predict(testDMatrix)
-    val testRowsRDD = sc.parallelize(
-      auxTestItr.toList.zipWithIndex, numWorkers).map {
-      case (instance: LabeledPoint, id: Int) =>
-        Row(id, instance.features, instance.label)
-    }
-    val trainingDF = buildTrainingDataframe()
-    val xgBoostModelWithDF = XGBoost.trainWithDataset(trainingDF, paramMap,
-      round = 5, nWorkers = numWorkers, useExternalMemory = false)
-    val testDF = trainingDF.sparkSession.createDataFrame(testRowsRDD, StructType(
-      Array(StructField("id", IntegerType), StructField("features", new VectorUDT),
-        StructField("label", DoubleType))))
-    val predResultsFromDF =
-      xgBoostModelWithDF.transform(testDF).collect().map(row => (row.getAs[Int]("id"),
-        row.getAs[mutable.WrappedArray[Float]]("prediction"))).toMap
-    for (i <- predResultFromSeq.indices) {
-      assert(predResultFromSeq(i).length === predResultsFromDF(i).length)
-      for (j <- predResultFromSeq(i).indices) {
-        assert(predResultFromSeq(i)(j) === predResultsFromDF(i)(j))
-      }
     }
   }
 
@@ -331,7 +139,6 @@ class XGBoostSuite extends FunSuite with BeforeAndAfter {
     }
     val trainingRDD = buildTrainingRDD()
     val testRDD = buildEmptyRDD()
-    import DataUtils._
     val tempDir = Files.createTempDirectory("xgboosttest-")
     val tempFile = Files.createTempFile(tempDir, "", "")
     val paramMap = List("eta" -> "1", "max_depth" -> "2", "silent" -> "0",
